+++ packages/composites/dns-record-check/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/dns-record-check-adapter
 
+## 1.3.30
+
+### Patch Changes
+
+- Updated dependencies [3dc13a0bc]
+  - @chainlink/ea-bootstrap@2.27.0
+  - @chainlink/ea-test-helpers@1.4.2
+  - @chainlink/dns-query-adapter@1.6.22
+
 ## 1.3.29
 
 ### Patch Changes