+++ packages/sources/icap/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/icap-adapter
 
+## 1.1.2
+
+### Patch Changes
+
+- 31af84e69: Bumped framework version
+- Updated dependencies [31af84e69]
+  - @chainlink/tp-adapter@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes