+++ packages/sources/tp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chainlink/tp-adapter
 
+## 1.1.2
+
+### Patch Changes
+
+- 31af84e69: Bumped framework version
+
 ## 1.1.1
 
 ### Patch Changes